api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx~=0.28
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydantic==2.5.0
pydantic-settings==2.0.3
supabase==2.15.1
PyJWT==2.8.0